scripts/requirements.txt
beautifulsoup4==4.11.2
python-dotenv==1.0.0
pymongo==4.3.3
urllib3==1.26.14